Move slsi_wifi from apps/ to external/
External libraries have to be moved to external/ folder. slsi_wifi belongs to this category, especially maintained by S.LSI.
