+++ apps/examples/artik_demo/artik_demo.c
@@ -39,7 +39,7 @@
 #include <dm/dm_error.h>
 #include <dm/dm_connectivity.h>
 
-#include <apps/netutils/wifi/slsi_wifi_api.h>
+#include <slsi_wifi/slsi_wifi_api.h>
 
 /****************************************************************************
  * Definitions
