+++ apps/examples/testcase/ta_tc/device_management/utc/utc_dm_main.c
@@ -28,7 +28,7 @@
 #include <dm/dm_connectivity.h>
 
 #include <apps/netutils/dhcpc.h>
-#include <apps/netutils/wifi/slsi_wifi_api.h>
+#include <slsi_wifi/slsi_wifi_api.h>
 
 #define NET_DEVNAME "wl1"
