+++ apps/examples/wakaama_client/wakaama_client.c
@@ -29,7 +29,7 @@
 
 #include <apps/netutils/dhcpc.h>
 
-#include <apps/netutils/wifi/slsi_wifi_api.h>
+#include <slsi_wifi/slsi_wifi_api.h>
 
 #define NET_DEVNAME "wl1"
