+++ external/iotivity/iotivity_1.2-rel/build_common/tizenrt/SConscript
@@ -78,11 +78,12 @@ if env.get('LOGGING'):
 
 dep_src_dir1 =  os.path.join(tizenrt_os_dir, 'include')
 dep_src_dir2 =  os.path.join(tizenrt_os_dir, 'include/tinyara')
+dep_src_dir3 =  os.path.join(tizenrt_os_dir, '../external/include')
 dep_lib_dir =  '/usr/lib/arm-none-eabi/include'
 
 
 # Add directories to search for header files and external libraries
-env.AppendUnique(CPPPATH = [ dep_src_dir1, dep_src_dir2 ])
+env.AppendUnique(CPPPATH = [ dep_src_dir1, dep_src_dir2, dep_src_dir3 ])
 env.AppendUnique(LIBPATH = [ dep_lib_dir ])
 
 print env.get('CPPPATH')
