+++ external/slsi_wifi/Make.defs
@@ -51,5 +51,15 @@
 ############################################################################
 
 ifeq ($(CONFIG_NETUTILS_WIFI),y)
-CONFIGURED_APPS += netutils/wifi
+
+CSRCS += slsi_wifi_api.c
+CSRCS += slsi_wifi_utils.c
+
+CFLAGS += -I$(TOPDIR)$(DELIM)../apps$(DELIM)wpa_supplicant$(DELIM)src
+CFLAGS += -I$(TOPDIR)$(DELIM)../apps$(DELIM)wpa_supplicant$(DELIM)src$(DELIM)utils
+CFLAGS += -I$(TOPDIR)$(DELIM)../apps$(DELIM)wpa_supplicant$(DELIM)src$(DELIM)common
+
+DEPPATH += --dep-path slsi_wifi
+VPATH += :slsi_wifi
+
 endif
