+++ framework/src/dm/arch/sidk_s5jt200/s5j_dm_connectivity.c
@@ -41,8 +41,8 @@
 #include <dm/dm_connectivity.h>
 
 #include <apps/netutils/dhcpc.h>
-#include <apps/netutils/wifi/slsi_wifi_api.h>
-#include <apps/netutils/wifi/slsi_wifi_utils.h>
+#include <slsi_wifi/slsi_wifi_api.h>
+#include <slsi_wifi/slsi_wifi_utils.h>
 #include <arpa/inet.h>
 
 #define DM_GET_INTERFACE_INFO 0
